Dynamic activities on the homepage

+++ database/migrations/2019_08_17_200000_create_activiteiten_table.php
@@ -22,7 +22,7 @@ class CreateActiviteitenTable extends Migration
             $table->time('start_uur')->default('14:00:00');
             $table->time('eind_uur')->default('17:00:00');
             $table->string('locatie')->default('Lokaal');
-            $table->integer('prijs')->unsigned()->default(0);
+            $table->float('prijs', 5, 2)->unsigned()->default(0);
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
         });
